+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 * Create and approve change role request for role that does  not exist
+* Removed ability to create sale for reversed pair
 
 ## 3.2.1
 ### Fixed