+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 
 * Manage account specific rule op frame
+* Cancel change role reviewable request
 
 ### Fixed
 * Create and approve change role request for role that does  not exist