+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Changelog
 
-## Unreleased
+## 3.2.0
+
 ### Added
 
 * Manage account specific rule op frame
 
 ### Fixed
+
 * Create and approve change role request for role that does  not exist
 
+
 ## 3.2.1
+
 ### Fixed 
 * Fail of histexists command if db has not been initialized